\name{print.gp}
\alias{print.gp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Spectral GP default print statement}
\description{
  This is the default print statement for a spectral GP object. If you need
     a list of everything that is part of a spectral GP object,
     use 'names()'. 
}
\references{For technical details, see Paciorek, C.J. 2006. Bayesian Smoothing of Irregularly-spaced Data Using Fourier Basis Functions. Harvard University Biostatistics Working Paper Series. Working Paper 49.  \url{http://www.bepress.com/harvardbiostat/paper49}}
  \author{Christopher Paciorek \email{paciorek@alumni.cmu.edu} }
\seealso{\code{\link{gp}},\code{\link{names.gp}}}

\keyword{models }% at least one, from doc/KEYWORDS
\keyword{smooth}% __ONLY ONE__keyword per line
\keyword{spatial}% __ONLY ONE__ keyword per line
