\name{spectralGP-generic}
\alias{add.blocks}
\alias{calc.variances}
\alias{change.param}
\alias{copy}
\alias{expand.gpgrid}
\alias{getgrid}
\alias{Gibbs.sample.coeff}
\alias{Hastings.coeff}
\alias{image_plot_info}
\alias{image_plot_plt}
\alias{logdensity}
\alias{propose.coeff}
\alias{simulate}
\alias{updateprocess}
\alias{zero.coeff}
\title{
  spectralGP generic functions
}
\description{
  These functions are generics; see the help files associated with the
  spectralGP methods.
}
\alias{propose.coeff}
\keyword{models }% at least one, from doc/KEYWORDS
\keyword{smooth}% __ONLY ONE__keyword per line
\keyword{spatial}% __ONLY ONE__ keyword per line

