\name{SpectralGP}
\alias{add.blocks}
\alias{calc.variances}
\alias{change.param}
\alias{copy}
\alias{expand.gpgrid}
\alias{getgrid}
\alias{Gibbs.sample.coeff}
\alias{Hastings.coeff}
\alias{image.plot.info}
\alias{image.plot.plt}
\alias{logdensity}
\alias{propose.coeff}
\alias{simulate}
\alias{updateprocess}
\alias{zero.coeff}
\title{
  SpectralGP internal and generic functions
}
\description{
  These functions are either generics (see help files associated with the
  spectralGP specific functions) or internal functions that are not intended for user use.
}
\keyword{internal}
