\name{print.gp}
\alias{print.gp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Spectral GP default print statement}
\description{
  This is the default print statement for a spectral GP object. If you need
     a list of everything that is part of a spectral GP object,
     use 'names()'. 
}
\references{Type 'citation("spectralGP")' for references.}
  \author{Christopher Paciorek \email{paciorek@alumni.cmu.edu} }
\seealso{\code{\link{gp}}, \code{\link{names.gp}}}

\keyword{models }% at least one, from doc/KEYWORDS
\keyword{smooth}% __ONLY ONE__keyword per line
\keyword{spatial}% __ONLY ONE__ keyword per line
